Split handler between websocket and REST
